🤖 Refactor: state 상태 분리 되돌리기 - 성능 측면에서 문제가 있을것이라고 판단
